refactor: move utils to separate files
